src/processor/requirements.txt
pyyaml
pygeohash
boto3
Pillow>=11.0.0
